Create file: mario.html again.
This time both the branch ('mario') and the file was created on github. 
After creating the branch and cloning it to my desktop. I could not see my new branch nor leifs' branch, only the 'master' branch.
So I added the file mario.html to see if my branch would show in my clone.